Alert Option Added

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -36,7 +36,24 @@ class Favorite extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite',
+                            'Are you sure you wish to Delete the Favorite dish ' + item.name+'?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name+' Not Deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'Ok',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false}
+                        );
+                    }   
                 }
             ];
